fix: remove dead ScanOptions API and fix numeric precision edges
Sweeping-review follow-ups (round 2):

- Remove ScanOptions: declared, re-exported and present in index.d.ts but
  never consumed by any API (scan options are set via YaraX setters,
  matching upstream yara-x Python binding methods). Deletes the type from
  types.rs, the re-export in lib.rs, and the regenerated index.d.ts. It was
  never usable, so no working code can depend on it.
- MatchData offsets: offset/length/contextMatchOffset were u32, silently
  wrapping on scans of files >= 4GiB (verified: a match at 2^32+50 was
  reported as 50). Bumped to i64 (napi maps i64 to JS number, so the d.ts
  surface is unchanged) and verified a >4GiB sparse-file scan reports the
  true offset (4294967346).
- get_variable_value integer boundary: '<= i64::MAX as f64' accepts 2^63,
  and 'value as i64' silently saturates it to i64::MAX. Changed to a strict
  '<' so only numbers exactly representable below 2^63 go Integer; larger
  values stay Float. (Not externally observable since yara-x coerces
  numerics for comparison, but removes the silent value corruption.)
- Non-issue confirmed: enable_feature/ignore_module/ban_module return
  &mut Self in yara-x 1.20 (no error surface), so the 'let _ =' in
  apply_compiler_options is correct — no silent failure to fix.
- Tests: match-offset/context-offset smoke test (113/113 passing; fmt and
  clippy clean).

+++ src/scanner.rs
@@ -221,14 +221,14 @@ impl YaraX {
         };
 
         let context_match_offset = if has_context {
-          Some(context_range.start as u32)
+          Some(context_range.start as i64)
         } else {
           None
         };
 
         matches_vec.push(MatchData {
-          offset: offset as u32,
-          length: length as u32,
+          offset: offset as i64,
+          length: length as i64,
           data: matched_data,
           identifier: pattern_id.clone(),
           context_data,
